#443: Moved the PassiveEquipmentCategoryStatsIncreaseParser to the model folder
 -
